hw2/requirements.txt
aiohttp==3.6.1
redis==3.3.8
python-dotenv==0.10.3
aiohttp_validate==1.1.0